Bug 869773: Add local mirror manifest.
The repo tool supports creating a local mirror using the init --mirror
option.  This downloads all of the repositories referenced in the manifest,
but does not check them out.  You can then reference this mirror by using
the --reference= option to repo init.

To work well the mirror needs to download as many common repositories as
possible.  This manifest represents an attempt to do this for our current
set of project manifests.  When there were conflicts I chose to use the
most commonly referenced source.

To use this manifest the config.sh script must be updated to support
creating and using mirrors.