Bug 952535 - Use Code Aurora's platform/system/core/ at b2g_ics_1.2
During build of FOTA packages, an updater-script is generated. Part of
its work is to extract new files from the ZIP update file and correctly
set the permissions. Those are set at build time of this script, via the
call of a specific binary. This binary relies on the
android_filesystem_config.h header file that lives inside
platform/system/core's include/private directory. In order to provide
update packages that works, we have to add the B2G binaries in this
header. Code Aurora's b2g_ics_1.2 is an ICS part of the repository that
has the correct patches ready for B2G, like this change. We make use of
this everywhere, to be able to produce updates for all devices we can.